The Directory Observer Moment
- Now Directory is using two observer patterns to check when it is
updated; and when a route is retrieved.

- UML has been update to reflect the fact the I have more of an idea of
how to implement observers in Directory

- The factory has been updated to also include a make_route() class.
This is because routes must specifically be made in the directory. It
will always make a new route to add to the route_list if the last object
created was a Waypoint (unless this specific route is already made in
route_list).